+++ src/pages/NotFound.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const NotFoundPage: React.FC = () => {
+  return (
+    <div>
+      <h1>404 - Error</h1>
+      <p>Sorry, Please try again</p>
+    </div>
+  );
+};
+
+export default NotFoundPage;
